requirements.txt
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
python-dotenv>=1.0.0
openai>=1.12.0 